+++ Naenio/Sources/LogIn/AppleLoginButton.swift
@@ -1,0 +1,18 @@
+//
+//  AppleLoginButton.swift
+//  Naenio
+//
+//  Created by 프라이빗 on 2022/07/10.
+//
+
+import SwiftUI
+import AuthenticationServices
+
+/// Sign in with Apple 기본 제공 버튼을 SwiftUI 사용하기 위한 래퍼 클래스
+final class AppleLoginButton: UIViewRepresentable {
+    func makeUIView(context: Context) -> ASAuthorizationAppleIDButton {
+        return ASAuthorizationAppleIDButton(type: .signIn, style: .whiteOutline)
+    }
+    
+    func updateUIView(_ uiView: ASAuthorizationAppleIDButton, context: Context) {}
+}
