--- Naenio/Sources/LogIn/LoginTestViewModel.swift
@@ -1,93 +1,0 @@
-//
-//  AppleLoginTestViewModels.swift
-//  Naenio
-//
-//  Created by 프라이빗 on 2022/07/10.
-//
-
-import SwiftUI
-import AuthenticationServices
-import KakaoSDKAuth
-
-class LoginTestViewModel: ObservableObject {
-    // Dependencies
-    let appleLoginManager: AppleLoginManager
-    let kakaoLoginManager: KakaoLoginManager
-    
-    // Published vars
-    /// 로그인 루틴 처리 상태를 표시
-    @Published var status: Status = .waiting
-    
-    // MARK: - For apple
-    func handleAppleLoginResult(result: Result<ASAuthorization, Error>) {
-        switch result {
-        case .success(let authResult):
-            requestAppleLoginToServer(with: authResult)
-        case .failure(let error):
-            status = .fail(with: error)
-        }
-    }
-    
-    private func requestAppleLoginToServer(with result: ASAuthorization) {
-        switch appleLoginManager.requestLoginToServer(with: result) {
-        case .success(let user):
-            print(user)
-            // Save(user)
-            
-            status = .done
-        case .failure(let error):
-            status = .fail(with: error)
-        }
-    }
-    
-    
-    // MARK: - For kakao
-    func handleKakaoLoginResult(result: Result<OAuthToken, Error>) {
-        switch result {
-        case .success(let authResult):
-            requestKakaoLoginToServer(with: authResult)
-        case .failure(let error):
-            status = .fail(with: error)
-        }
-    }
-    
-    private func requestKakaoLoginToServer(with result: OAuthToken) {
-        switch kakaoLoginManager.requestLoginToServer(with: result) {
-        case .success(let user):
-            print(user)
-            // Save(user)
-            
-            status = .done
-        case .failure(let error):
-            status = .fail(with: error)
-        }
-    }
-    
-    init(_ appleLoginManager: AppleLoginManager = AppleLoginManager(),
-         _ kakaoLoginManager: KakaoLoginManager = KakaoLoginManager()) {
-        self.appleLoginManager = appleLoginManager
-        self.kakaoLoginManager = kakaoLoginManager
-    }
-}
-
-extension LoginTestViewModel {
-    enum Status {
-        case waiting
-        case inProgress
-        case done
-        case fail(with: Error)
-        
-        var description: String {
-            switch self {
-            case .waiting:
-                return "Waiting"
-            case .inProgress:
-                return "In progres..."
-            case .done:
-                return "Successfully done"
-            case .fail(let error):
-                return "Failed with error: \(error.localizedDescription)"
-            }
-        }
-    }
-}
