+++ Naenio/Utils/Models/LoginRequestInformation.swift
